fixed my issue with company ids not populating

+++ public/js/dashboard.js
@@ -29,7 +29,9 @@ function createHackathon(hackathonData) {
     method: "post",
     data: hackathonData,
   })
-    .then(function(hackathonData) {})
+    .then(function() {
+      location.reload();
+    })
     .catch(handleHackathonErr);
 }
 
@@ -40,24 +42,46 @@ function handleHackathonErr(err) {
 
 $(".hackathonDelete").on("click", function(event) {
   var id = $(this).data("id");
-  // Send the DELETE request.
-  $.ajax("/api/hackathons/" + id, {
-    type: "DELETE",
-  }).then(function() {
-    console.log("deleted hackathon ", id);
-    // Reload the page to get the updated list
-    location.reload();
-  });
-});
-
-$(".hackathonJoin").on("click", function(event) {
-    var id = $(this).data("id");
-    // Send the POST request.
-    $.ajax("/api/users/join-hackathon/" + id, {
-      type: "POST",
+  if (confirm("Are you sure you want to delete this Hackathon?")) {
+    // Send the DELETE request.
+    $.ajax("/api/hackathons/" + id, {
+      type: "DELETE",
     }).then(function() {
-      console.log("joined hackathon ", id);
+      console.log("deleted hackathon ", id);
       // Reload the page to get the updated list
       location.reload();
     });
-  });
+  }
+});
+
+$(".hackathonJoin").on("click", function(event) {
+  var id = $(this).data("id");
+  // Send the POST request.
+  $.ajax("/api/users/join-hackathon/" + id, {
+    type: "POST",
+  })
+    .then(function() {
+      console.log("joined hackathon ", id);
+      // Reload the page to get the updated list
+      location.reload();
+    })
+    .catch(function(err) {
+      alert("You have already joined that hackathon");
+    });
+});
+
+$(".hackathonUnjoin").on("click", function(event) {
+  var id = $(this).data("id");
+  // Send the DELETE request.
+  $.ajax("/api/hackathons/unjoin/" + id, {
+    type: "DELETE",
+  })
+    .then(function() {
+      console.log("unjoined hackathon ", id);
+      // Reload the page to get the updated list
+      location.reload();
+    })
+    .catch(function(err) {
+      alert(err);
+    });
+});
